git-svn-id: http://localhost:8080/svn/RsdnNntp/trunk@555 259efaa1-37b8-6b43-839f-2492ada40869

+++ NNTP/Commands/Next.cs
@@ -39,7 +39,7 @@ namespace Rsdn.Nntp.Commands
 
 			NewsArticle article = session.DataProvider.GetNextArticle(session.currentArticle, session.currentGroup);
 
-			session.currentArticle = (int)article.MessageNumbers[session.currentGroup];
+			session.currentArticle = article.MessageNumbers[session.currentGroup];
 
 			return new Response(NntpResponse.ArticleNothingRetrivied, null,
 				article.MessageNumbers[session.currentGroup], article["Message-ID"]);
